# app/services/appstore-api/swagger/swagger.yaml
basePath: /
consumes:
    - application/json
    - application/yaml
...
host: localhost:3000
info:
    contact:
        email: jseals@renci.org
        name: Joshua Seals
    description: |-
        The purpose of this application is to provide a
        pluggable REST interface to the helxplatform that
        will act as orchestrator among varying microservices.
    license:
        name: MIT
        url: http://opensource.org/licenses/MIT
    termsOfService: there are no TOS at this moment, use at your own risk we take no responsibility
    title: GopherHelx API.
    version: 1.0.0
paths:
...
    /app/new:
        post:
            description: |-
                Expects an Application definition
                Responds Success or Error Json Response
            operationId: newApplication
            parameters:
                - description: the name of the application
                  example: scipy
                  in: query
                  name: app_name
                  required: true
                  type: string
                  x-go-name: AppName
                - description: the image of the application
                  example: helxplatform/scipy
                  in: query
                  name: image
                  required: true
                  type: string
                  x-go-name: Image
                - description: the port for the application
                  example: 8888
                  format: int64
                  in: query
                  name: port
                  required: true
                  type: integer
                  x-go-name: Port
            responses:
                "200":
                    $ref: '#/responses/newApplication'
            tags:
                - apps
...
produces:
    - application/json
responses:
...
    newApplication:
        description: ""
        schema:
            properties:
                Success:
                    description: Success message with appID
                    example: "117"
                    type: string
            type: object
...
schemes:
    - http
swagger: "2.0"